CaesarSQLDB/requirements.txt
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SocketIO==5.3.5
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient
pip==23.2.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
setuptools==68.0.0
simple-websocket==0.10.1
urllib3==2.0.4
Werkzeug==2.3.6
wheel==0.38.4
wsproto==1.2.0
zipp==3.16.2
